Add class to manage phabricator daemon launcher: phd
We provide a wrapper (as an init script) around phd to use it as a
service in puppet. Also, since phd needs a correct name resolution for
the base URI, we also add a host resource in phabricator's main class
to map the fully qualified domain name to the local host.